refactor(tmux): extract src/tmux/ into par-term-tmux subcrate (#184)
Move ~2,474 lines of tmux control mode integration code from src/tmux/
into a new par-term-tmux workspace subcrate.

- Create par-term-tmux/Cargo.toml with deps: par-term-config,
  par-term-emu-core-rust, winit, chrono, hostname, regex, log
- Move all 8 source files from src/tmux/ to par-term-tmux/src/
- Create par-term-tmux/src/lib.rs (replaces src/tmux/mod.rs)
- Update intra-module imports: super:: → crate::
- Replace crate::debug_*! macros with log::info!/error!/trace!
- Add par-term-tmux to workspace members in root Cargo.toml
- Add par-term-tmux as dependency in main crate
- Replace `pub mod tmux` with `pub use par_term_tmux as tmux` in
  src/lib.rs — zero call-site changes required

All 40 unit tests in par-term-tmux pass. Full workspace build and
test suite pass with no regressions.

Closes #184
