+++ src/lib.rs
@@ -70,7 +70,7 @@ pub mod tab_bar_ui;
 pub mod terminal;
 pub mod text_shaper;
 pub mod themes;
-pub mod tmux;
+pub use par_term_tmux as tmux;
 pub mod tmux_session_picker_ui;
 pub mod tmux_status_bar_ui;
 pub mod update_checker;
